Clean before Fix
